% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_events.R
\name{group_events}
\alias{group_events}
\title{Group events of a common term group}
\usage{
group_events(
  .data,
  t_groups,
  n_event,
  n_arm = NULL,
  event_name,
  study_id,
  src = "src",
  method = c("conservative", "integrative"),
  add_zero = FALSE
)
}
\arguments{
\item{.data}{An event data.frame (see details)}

\item{t_groups}{Named list of terms. Names are terms, contents are event names (see description).}

\item{n_event}{Character vector, column with number of events in study arms (e.g. placebo...).}

\item{n_arm}{Optional, character vector, column with number of patients in study arms.}

\item{event_name}{Character string, column of events name.}

\item{study_id}{Character string, column of main study ID}

\item{src}{Character vector, column with source sorting \code{\link{sort_sources}}.}

\item{method}{"integrative" or "conservative" (see details).}

\item{add_zero}{Do you wish to count O events for study(ies) without any events of a term? (see description)}
}
\value{
Special \code{attributes} of the output data.table are
\itemize{
\item \code{group_method} (conservative or integrative)
\item \code{group_term}
\item \code{event}
\item \code{n}
} They are used in subsequent functions to create a single dataset when there are multiple arms.
}
\description{
Allows for prioritization of event data, aggregating multiples terms under a group term (see integrative and conservative methods).
}
\details{
See \code{vignette("basic_workflow", package = "safetymeta")} for details on
"event" and "term" definition.
With this function, you can group multiple events corresponding to a term.
A t_groups list provides one or more \code{terms} as its name(s).
\strong{Integrative} method means counts for each event of a \code{term}
will be summed up together, while \strong{conservative} means only the event
with the highest count will be selected.
\code{add_zero} is used when an event is not reported for a study
(e.g., the event line is NOT in the dataset).
The principle is that since all serious events must be reported,
the absence of reporting is equal to reporting 0 events.
This is not true when safety data is not available at all
(i.e., when there is no data for such event in any available sources).
Thus, these specific studies must be \strong{excluded} prior to using \code{group_events}.
Some total counts are not homogeneous for a given study, depending on the reaction (notably found for all cause mortality reaction). The maximum count will be retained for no line study(ies), (when \code{add_zero} performs).
\strong{\code{group_events} assumes events in \code{.data[["events"]]} are LOWER cased.}
}
\examples{


# You are very much advised to check string case over t_groups
# and .data[["event_name"]] col. All should be lower cased +++

source_order <- c("src1", "src2")

# set src before using group_events
e_data <-
  e_data |>
  sort_sources(
    method = c("source_list"),
    source_name = "source",
    source_list_order = source_order
  )

# The integrative method
 e_data |>
  group_events(
    t_groups = t_groups,
    n_event = "n_event_arm1",
    n_arm     = "n_arm1",
    event_name = "event_name",
    study_id   = "study_id",
    method = "integrative"
  )

# Contrasting with the conservative method

 e_data |>
  group_events(
    t_groups = t_groups,
    n_event = "n_event_arm1",
    n_arm     = "n_arm1",
    event_name = "event_name",
    study_id   = "study_id",
    method = "conservative"
  )
}
\keyword{meta-analysis}
