% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetymeta-package.R
\docType{package}
\name{safetymeta-package}
\alias{safetymeta}
\alias{safetymeta-package}
\title{safetymeta: Conduct safety meta-analysis of study-level data}
\description{
Guidance through data management steps of safety events, obtained from different sources (manuscripts, clinical trial registration platforms), grouping into safety terms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmacologie-caen.github.io/safetymeta/}
}

}
\author{
\strong{Maintainer}: Charles Dolladille \email{cdolladille@hotmail.com} (\href{https://orcid.org/0000-0003-0449-6261}{ORCID})

}
\keyword{internal}
