% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_sources.R
\name{sort_sources}
\alias{sort_sources}
\title{Define source order}
\usage{
sort_sources(
  .data,
  method = c("source_list", "max_n"),
  source_name = NULL,
  source_list_order = NULL,
  study_id = NULL,
  study_n = NULL
)
}
\arguments{
\item{.data}{A study- or event-level dataset @seealso \link{e_data}}

\item{method}{Character string, one of "source_list", "max_n". See description.}

\item{source_name}{Optional, character string, a column with sources}

\item{source_list_order}{Optional, character vector, order of sources.}

\item{study_id}{Optional, study identifier.}

\item{study_n}{Optional, character string, a column with numbers (of patients).}
}
\value{
The original .data dataset, with an additional \code{src} numeric column.
}
\description{
Prioritize over multiple sources, according to some criteria
}
\details{
When multiples sources report safety data for the same event,
they should be prioritized among each other, in order to retain the source
with the highest priority for each event.
\itemize{
\item If source sorting has been decided use the \code{source_list} method.
With this method, your dataset should have a column with the \code{source_name},
and you provide the order in the \code{source_list_order} argument.
The first element of \code{source_list_order} is the highest priority
source, the second element is the next highest priority, and so on.
\item If you want to sort sources according to highest number of patients
in a study_arm (or any other number), use method \code{max_n}, and provide
\code{study_id} and \code{study_n} arguments.
}
}
\examples{

# Source order is already known, then use method "source_list"

 source_order <- c("src1", "src2")

 e_data |>
   sort_sources(
     method = c("source_list"),
     source_name = "source",
     source_list_order = source_order
   )

 # Sources with higher total number in arm1 have highest priority

   e_data_maxn <-
      e_data |>
        dplyr::select(study_id, n_arm1) |>
        dplyr::add_row(
          study_id = "NCT1",
          n_arm1   = 30
        )

  e_data_maxn |>
  sort_sources(
    method = c("max_n"),
    study_n = "n_arm1",
    study_id = "study_id"
    )
}
